test(app): move the fault-message wording tests to their own file
Pure move, no assertion changed: the same 30 tests run before and after.
ChannelFaultIntegrationTests answers when a fault is reported at all; these
answer what the resulting message tells the user to go and do, and the file
was at the 600-line lint ceiling with more of the second kind to come.
